+++ default.nix
@@ -1,19 +1,15 @@
-{ mkDerivation, aeson, base, binary, bytestring, cereal
-, constraints, deepseq, hashable, stdenv, store, tasty, tasty-hunit
-, tasty-quickcheck
+let
+_pkgs0 = import ((import <nixpkgs> {}).fetchFromGitHub {
+  owner = "NixOS";
+  repo = "nixpkgs";
+  rev = "4aadb9beb345898c7f7b06090b163e7f3debea5a";
+  sha256 = "1dka2knhi8pfb83iwph7chldys1df95dc3z2v38cqzy4m06qjir9";
+}) {};
+
+in
+{ pkgs ? _pkgs0
+, compiler ? "ghc801"
 }:
-mkDerivation {
-  pname = "safe-money";
-  version = "0.1";
-  src = ./.;
-  libraryHaskellDepends = [
-    aeson base binary cereal constraints deepseq hashable store
-  ];
-  testHaskellDepends = [
-    aeson base binary bytestring cereal constraints deepseq hashable
-    store tasty tasty-hunit tasty-quickcheck
-  ];
-  homepage = "https://github.com/k0001/safe-money";
-  description = "Type-safe and lossless encoding and manipulation of money, world currencies and precious metals";
-  license = stdenv.lib.licenses.bsd3;
-}
+
+let drv = pkgs.haskell.packages.${compiler}.callPackage (import ./pkg.nix) {};
+in if pkgs.lib.inNixShell then drv.env else drv