+++ changelog.md
@@ -1,3 +1,8 @@
+# Version HEAD
+
+* Backwards compatibility with GHC 7.10.
+
+
 # Version 0.1
 
 * Initial release.