+++ changelog.md
@@ -1,5 +1,8 @@
 # Version HEAD
 
+* Cabal flags are now manual (`aeson`, `binary`, `bytes`, `cereal`, `deepseq`,
+  `hashable`).
+
 * Backwards compatibility with GHC 7.10.
 
 * Fix `Store` instances and test them.